fix(tools): harden native workspace discovery
